requirements.txt
numpy
opencv-python
supervision
ultralytics
pure-python-adb
scrcpy-client
matplotlib
